+++ src/components/app/App.js
@@ -1,42 +1,29 @@
-import { useState } from "react";
-import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ComicsList from "../comicsList/ComicsList";
-import AppBanner from "../appBanner/AppBanner";
-import ErrorBoundary from "../errorBoundry/ErrorBoundry";
+import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
 
-import decoration from '../../resources/img/vision.png';
+import {MainPage, ComicsPage} from '../pages';
+//якщо файл
+// не вказаний вебпак шукає в файлі індекс.джс, 
+//якщо такого файлу нема - помилка
+import AppHeader from "../appHeader/AppHeader";
 
 const App = () => {
-	
-	const [selectedChar, setChar] = useState(null);
-
-	const onCharSelected = (id) => {
-		setChar(id);
-	}
 
 	return (
-		<div className="app">
-			<AppHeader/>
-			<main>
-				{/* <ErrorBoundary>
-					<RandomChar/>
-				</ErrorBoundary>
-				<div className="char__content">
-					<ErrorBoundary>
-						<CharList onCharSelected={onCharSelected}/>
-					</ErrorBoundary>
-					<ErrorBoundary>
-						<CharInfo charId={selectedChar}/>
-					</ErrorBoundary>
-				</div>
-				<img className="bg-decoration" src={decoration} alt="vision"/> */}
-				<AppBanner/>
-				<ComicsList/>
-			</main>
-		</div>
+		<Router>
+			<div className="app">
+				<AppHeader/>
+				<main>
+					<Switch>
+						<Route exact path="/">
+							<MainPage/>
+						</Route>
+						<Route exact path="/comics">
+							<ComicsPage/>
+						</Route>
+					</Switch>
+				</main>
+			</div>
+		</Router>
 	)
 }
